Updated framework code structure
