+++ test_purchase_item.py
@@ -1,0 +1,6 @@
+from base_test import BaseTest
+
+
+class PuchaseItem(BaseTest):
+    def testFirstAutomationTest(self):
+        self.ui_workflow.login()
